+++ source/main/include/cbase/private/c_float_inline.h
@@ -177,7 +177,7 @@ namespace math
     /**
      * Exponent. Returns the Exponent part of the floating point value, without the bias
      */
-    inline s32 exponent(f32 f) { return exponentBinary(f) - FF_EXPONENT_BIAS; }
+    inline s32 exponent(f32 f) { return exponentBinary(f) - (s32)FF_EXPONENT_BIAS; }
 
     /**
      * SignMask. Returns the floating point value masked with the sign mask.
@@ -286,7 +286,7 @@ namespace math
      */
     inline f32 minimumIncrement(f32 inFloat)
     {
-        u32 exponent = exponentBinary(inFloat) - FF_FRACTION_BITS;
+        u32 exponent = (u32)(exponentBinary(inFloat) - FF_FRACTION_BITS);
 
         if (exponent > FF_EXPONENT_MASK >> FF_GET_EXPONENT_SHIFT)
             exponent = FF_EXPONENT_MASK >> FF_GET_EXPONENT_SHIFT;
